src/main/java/edu/depaul/ticketselling/management/service/PurchaseService.java
package edu.depaul.ticketselling.management.service;

import edu.depaul.ticketselling.management.interfaces.IPurchaseService;
import edu.depaul.ticketselling.management.model.Purchase;
import edu.depaul.ticketselling.management.repository.PurchaseRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.time.LocalDate;
import java.util.List;
import java.util.Optional;
import java.util.stream.Collectors;
import java.util.stream.StreamSupport;

@Service
public class PurchaseService implements IPurchaseService {

    private final PurchaseRepository purchaseRepository;

    @Autowired
    public PurchaseService(PurchaseRepository purchaseRepository) {
        this.purchaseRepository = purchaseRepository;
    }

    public Purchase save(Purchase purchase) {
        return purchaseRepository.save(purchase);
    }

    public List<Purchase> findAll() {
        return purchaseRepository.findAll();
    }

    public Purchase findById(Long id) {
        Optional<Purchase> purchase = purchaseRepository.findById(id);
        return purchase.orElse(null);
    }

    public void deleteById(Long id) {
        purchaseRepository.deleteById(id);
    }
}
